refactor(vue): move from jsdoc to ts
