+++ packages/vue/src/app.ts
@@ -1,0 +1,9 @@
+import { App } from './apiCreateApp';
+import { onUnmounted, $unison } from '@unisonjs/core';
+
+export const AppProvider = $unison<{ app: App; children: React.ReactNode }>((props) => {
+  onUnmounted(() => {
+    props.app.unmount();
+  });
+  return () => props.children;
+}, 'AppProvider');
