+++ packages/vue/src/reactivity/arrayInstrumentations.ts
@@ -1,6 +1,6 @@
 import { TrackOpTypes } from './constants'
 import { isProxy, isShallow, toRaw, toReactive } from './reactive'
-import { endBatch, pauseTracking, resetTracking, startBatch, ARRAY_ITERATE_KEY, track } from '@briddge/core';
+import { endBatch, pauseTracking, resetTracking, startBatch, ARRAY_ITERATE_KEY, track } from '@unisonjs/core';
 import { isArray } from '@vue/shared'
 
 /**
